pyproject.toml
[project]
name = "django-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.8.1",
    "brotli>=1.1.0",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "colorama>=0.4.6",
    "django==2.2",
    "django-utils-six>=2.0",
    "djangorestframework==3.12.4",
    "djongo>=1.2.31",
    "dnspython>=2.7.0",
    "filelock>=3.18.0",
    "fsspec>=2025.3.2",
    "huggingface-hub>=0.30.2",
    "idna>=3.10",
    "jinja2>=3.1.6",
    "markupsafe>=3.0.2",
    "mpmath>=1.3.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "packaging>=25.0",
    "pillow>=11.2.1",
    "psycopg>=3.2.7",
    "psycopg-binary>=3.2.7",
    "pymongo<4.0",
    "python-decouple>=3.8",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "safetensors>=0.5.3",
    "setuptools>=80.3.1",
    "six>=1.17.0",
    "sqlparse==0.2.4",
    "sympy>=1.14.0",
    "tokenizers>=0.21.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "typing-extensions>=4.13.2",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "waitress>=3.0.2",
    "whitenoise>=6.9.0",
]
